Made changes to drawer UI
