+++ lib/provider/disability_notifier.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/widgets.dart';
+
+class DisabilityNotifier with ChangeNotifier {
+  bool isDisabilityEnabled = false;
+
+  void setDisabilityEnabled(bool disability) {
+    isDisabilityEnabled = disability;
+    notifyListeners();
+  }
+}
